Better Practices & Looping
Upon some research discovered that my handling of the global variables to start the game wasn't perfect practice. Changed it to better fit what I've seen others use.

Also now the game loops until the player decides if they want to quit or not. With the new looping changed some text and spacing to better fit.